поправил детект

+++ app/src/main/java/ru/practicum/android/diploma/data/dto/VacancySearchResponse.kt
@@ -4,7 +4,7 @@ import ru.practicum.android.diploma.data.dto.model.VacancyItemDto
 
 class VacancySearchResponse(
     val items: List<VacancyItemDto>, // Список вакансий
-    val found: Long, // Количество найденных вакансий
+    val found: Int, // Количество найденных вакансий
     val pages: Int, // Всего страниц (по 20 вакансий на страницу)
     val page: Int // Номер текущей страницы
 ) : Response()
